Chris Mason's patch from 20160719110534.GA96833@clm-mbp.thefacebook.com
On Mon, Jul 11, 2016 at 11:28:01AM +0530, Chandan Rajendra wrote:
> Hi Chris,
>
> I am able to reproduce the issue with the 'short-write' program. But before
> the call trace associated with btrfs_destroy_inode(), I see the following call
> trace ...
>
> ------------[ cut here ]------------
> WARNING: CPU: 2 PID: 2311 at /home/chandan/repos/linux/fs/btrfs/extent-tree.c:4303
+btrfs_free_reserved_data_space_noquota+0xe8/0x100

[ ... ]

Ok, the problem is in how we're dealing with the offset into the sector when
we fail. The dirty_sectors variable already has this accounted in it, so
this patch fixes it for me.  I ran overnight, but I'll let it go for a few
days just to make sure: