sd: wait for slow devices on shutdown path
Wait 1s during suspend/shutdown for the device to settle after
we issue the STOP command.

Otherwise we race ATA SSDs to powerdown, possibly causing damage to
FLASH/data and even bricking the device.

This is an experimental patch, there are likely better ways of doing
this that don't punish non-SSDs.

Signed-off-by: Henrique de Moraes Holschuh <hmh@hmh.eng.br>