btrfs: create a mount option for dax
This sets S_DAX in inode->i_flags, which can be used with
IS_DAX().

The dax option is restricted to non multi-device mounts.
dax interacts with the device directly instead of using bio, so
all bio-hooks which we use for multi-device cannot be performed
here. While regular read/writes could be manipulated with
RAID0/1, mmap() is still an issue.

Auto-setting free space tree, because dealing with free space
inode (specifically readpages) is a nightmare.
Auto-setting nodatasum because we don't get callback for writing
checksums after mmap()s.
Deny compression because it does not go with direct I/O.

Store the dax_device in fs_info which will be used in iomap code.

I am aware of the push to directory-based flags for dax. Until, that
code is in the kernel, we will work with mount flags.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>